Allow lit.util.which() to find executables with extension.
LLVM usually specifies executable names without extension so that
they work portably across platforms.  Occasionally, if you're writing
something platform specific, you might want to specify the extension.
For example, you might want to write a test that invokes a batch file.
It's awkward to say foo when the file is really called foo.bat, but
in this case lit.util.which() currently won't find it because it will
construct the filename foo.bat.bat.